+++ src/Traits/Has_Blocks.php
@@ -2,13 +2,14 @@
 
 namespace BernskioldMedia\WP\Block_Plugin_Support\Traits;
 
-use Exception;
-
 /**
  * Trait Has_Blocks
  *
  * Designed to extend a core plugin file that is a plugin
  * that provides blocks for output.
+ *
+ * @property array $conditional_blocks An array of blocks that are only loaded if the corresponding class name exists. blockname => classname.
+ * @property array $dynamic_blocks An array of blocks that are dynamic, meaning they use PHP to load on the frontend. blockname => classname.
  *
  * @package BernskioldMedia\WP\Block_Plugin_Support\Traits
  */
@@ -20,152 +21,40 @@ trait Has_Blocks {
 	protected array $blocks = [];
 
 	/**
-	 * Packages that are loaded as dependencies for the blocks globally,
-	 * unless the block specifies its own dependency list.
-	 *
-	 * @var string[]
-	 */
-	protected static array $default_block_dependencies = [
-		'wp-blocks',
-		'wp-components',
-		'wp-element',
-		'wp-i18n',
-		'wp-block-editor',
-	];
-
-	/**
-	 * Place the logic where you add blocks in here.
-	 */
-	abstract public function blocks(): void;
-
-	/**
-	 * Run this function when you initialize the plugin
-	 * so that blocks are registered.
+	 * Get all of the blocks in the blocks folder and register them.
 	 */
 	protected function load_blocks(): void {
-		if ( ! property_exists( static::class, 'block_prefix' ) ) {
-			throw new Exception( 'block_prefix is not set.' );
-		}
+		$blocks = glob( static::get_path() . 'blocks/*' );
 
-		add_action( 'plugins_loaded', [ $this, 'blocks' ], static::get_load_priority() );
-		add_action( 'init', [ $this, 'register_blocks' ], static::get_load_priority() );
-		add_action( 'render_block_data', [ $this, 'add_block_name' ], 10, 1 );
-		add_action( 'enqueue_block_editor_assets', [ $this, 'load_block_assets' ], static::get_block_assets_load_priority() );
+		$this->blocks = apply_filters( static::$slug . '_blocks', $blocks );
 	}
 
-	/**
-	 * Add a block.
-	 *
-	 * @return static
-	 */
-	protected function add_block( string $name, array $args = [] ) {
-		$this->blocks = array_merge( $this->blocks, [
-			$name => $args,
-		] );
-
-		return $this;
-	}
-
-	/**
-	 * Ad a block given that a class exists.
-	 *
-	 * @return static
-	 */
-	protected function add_block_if( string $class, string $name, array $args = [] ) {
-		if ( class_exists( $class ) ) {
-			$this->add_block( $name, $args );
-		}
-
-		return $this;
-	}
-
-	/**
-	 * Remove a block.
-	 *
-	 * @return static
-	 */
-	protected function remove_block( string $name ) {
-		unset( $this->blocks[ $name ] );
-
-		return $this;
-	}
-
-	/**
-	 * Load all block assets.
-	 */
-	public function load_block_assets(): void {
-		foreach ( $this->get_blocks() as $block_name => $args ) {
-			$dependencies = $args['script_dependencies'] ?? static::get_block_dependencies();
-
-			wp_enqueue_script( static::$block_prefix . '-' . $block_name, static::get_url( 'dist/' . $block_name . '.js' ), $dependencies, static::get_version(), false );
-			wp_set_script_translations( static::$block_prefix . '-' . $block_name, static::get_textdomain(), static::get_path( 'languages/' ) );
-		}
-	}
-
-	/**
-	 * Register the blocks
-	 */
 	public function register_blocks(): void {
-		foreach ( $this->get_blocks() as $block_name => $args ) {
-			register_block_type( static::$block_prefix . '/' . $block_name, $args );
+		foreach ( $this->blocks as $directory ) {
+			$parts = explode( '/', $directory );
+			$name  = end( $parts );
+
+			// If this is a conditional block and condition not met, then skip loading.
+			if ( isset( static::$conditional_blocks[ $name ] ) && ! class_exists( static::$conditional_blocks[ $name ] ) ) {
+				continue;
+			}
+
+			// Get the asset file data.
+			$asset_meta = include static::get_path() . 'dist/' . $name . '.asset.php';
+
+			// Register the script with WordPress.
+			wp_register_script( 'bm-block-' . $name, static::get_url( 'dist/' . $name . '.js' ), $asset_meta['dependencies'], $asset_meta['version'], true );
+
+			// Register the block. Dynamic blocks get their callback.
+			if ( isset( static::$dynamic_blocks[ $name ] ) ) {
+				register_block_type( $directory, [
+					'render_callback' => [ static::$dynamic_blocks[ $name ], 'render' ],
+				] );
+			} else {
+				register_block_type( $directory );
+			}
 		}
 	}
-
-	/**
-	 * Add the block name to the block attributes
-	 * so that we have it when rendering blocks via PHP.
-	 */
-	public function add_block_name( array $block ): array {
-		$block['attrs']['_name'] = $block['blockName'];
-
-		return $block;
-	}
-
-	/**
-	 * Get Blocks
-	 */
-	public function get_blocks(): array {
-		return $this->blocks;
-	}
-
-	/**
-	 * Control the priority of when this plugin hooks into
-	 * the init and plugins_loaded hooks.
-	 **/
-	public static function get_load_priority(): int {
-		if ( property_exists( static::class, 'load_priority' ) ) {
-			return static::$load_priority;
-		}
-
-		return 99;
-	}
-
-	/**
-	 * The priority of the block editor asset enqueue hook.
-	 */
-	public static function get_block_assets_load_priority(): int {
-		if ( property_exists( static::class, 'block_assets_load_priority' ) ) {
-			return static::$block_assets_load_priority;
-		}
-
-		return 99;
-	}
-
-	/**
-	 * Get the default block dependencies.
-	 */
-	public static function get_block_dependencies(): array {
-		if ( property_exists( static::class, 'block_dependencies' ) ) {
-			return array_merge( static::$block_dependencies, static::$default_block_dependencies );
-		}
-
-		return static::$default_block_dependencies;
-	}
-
-	/**
-	 * Get the Plugin's Version
-	 */
-	abstract public static function get_version(): string;
 
 	/**
 	 * Get the URL to the plugin folder, or the specified
@@ -178,10 +67,4 @@ trait Has_Blocks {
 	 * file relative to the plugin folder home.
 	 */
 	abstract public static function get_path( string $file = '' ): string;
-
-	/**
-	 * Get Plugin Textdomain
-	 */
-	abstract public static function get_textdomain(): string;
-
 }